+++ src/services/FCSServices/FCSServices.tsx
@@ -60,12 +60,12 @@ class FCSServices {
   }
 
   logicleMarkTransformer(
-    data: number[] | Int32Array,
+    data: number[] | Float32Array,
     rangeBegin?: number,
     rangeEnd?: number
-  ): Int32Array {
+  ): Float32Array {
     const logicle = new MarkLogicle(rangeBegin, rangeEnd);
-    return new Int32Array(data.map((e) => logicle.scale(e)));
+    return new Float32Array(data.map((e) => logicle.scale(e)));
   }
 
   logicleInverseMarkTransformer(
